refactor(server): extract multi-process guard to utils/multiprocess_guard.py
Task B of docs/current/server-py-module-split.md. Pure extraction,
verbatim move verified byte-identical against git history
(_detect_requested_worker_count + _warn_if_multi_process_requested).
New module owns its own structlog logger, matching the core/utils
pattern. server.py's lifespan call site is unchanged; only the import
moved. Functional check confirms detection + warning fire from the new
location and server.py resolves the same function object.
